Add EventTrend component, fix by_period to include gaps
Before we were only including a DatePeriodContainer for each period with
events, we now correctly take gaps into account and return empty Views
if no events occurred in that period. We also ensure
DatePeriodContainers are in order now, whereas before we were relying on
the input being sorted already (which it should be).

This is noticeably slower, but correct at least.